update default probe offset Z axis
now I'm using a ruthex 0.4 nozzle instead of the default one, and
I'll probably stick with those for the time being. however, this needs
to be calibrated any time I change the nozzle, because the positioning
of bltouch is not precise, and the nozzle torque isn't either. but I'll
still update the config because I like to reset default setting
sometimes.